+++ server/db/update_scripts/34_update_v7.1.0_v7.2.0.sql
@@ -1,0 +1,7 @@
+-- set DB version
+UPDATE version
+SET version = 'v7.2.0';
+
+-- change the config column in formActions to json
+ALTER TABLE formActions
+MODIFY COLUMN config JSON;